+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Authentication Zero 4.0.0 ##
+
+* Remove system tests
+* Use native rate_limit for lockable
+* Copy web_authn_controller.js instead of depend on stimulus-web-authn
+
 ## Authentication Zero 3.0.2 ##
 
 * Fix bug where token is not expired/invalid